pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.2",
    "numpy>=2.3.3",
    "pysolar>=0.13",
    "scipy>=1.16.2",
    "flask-migrate>=4.1.0",
    "flask-session>=0.8.0",
    "requests>=2.32.5",
    "cryptography>=46.0.1",
    "pyjwt>=2.10.1",
    "psycopg2-binary>=2.9.10",
]
